Address comments

+++ contracts/l1/L1Escrow.sol
@@ -9,7 +9,7 @@ interface ApproveLike {
 contract L1Escrow {
 
   // --- Auth ---
-  mapping (address => uint256) public wards;
+  mapping (address => uint8) public wards;
   function rely(address usr) external auth {
     wards[usr] = 1;
     emit Rely(usr);
